+++ madmom/utils/__init__.py
@@ -12,6 +12,7 @@ from __future__ import absolute_import, division, print_function
 
 import argparse
 import contextlib
+
 import numpy as np